refactor(node): tighten intent definition flow
