+++ packages/node/src/cli/intentCommands.ts
@@ -15,13 +15,24 @@ import { getIntentPaths, getIntentResultStepName, intentCatalog } from './intent
 import type { IntentFieldKind, IntentFieldSpec } from './intentFields.ts'
 import { inferIntentFieldKind } from './intentFields.ts'
 import type { IntentInputPolicy } from './intentInputPolicy.ts'
+import type {
+  IntentCommandDefinition,
+  IntentFileCommandDefinition,
+  IntentNoInputCommandDefinition,
+  IntentSingleStepExecutionDefinition,
+} from './intentRuntime.ts'
 import {
   createIntentOption,
   GeneratedBundledFileIntentCommand,
   GeneratedNoInputIntentCommand,
   GeneratedStandardFileIntentCommand,
   GeneratedWatchableFileIntentCommand,
+  getIntentOptionDefinitions,
 } from './intentRuntime.ts'
+import {
+  imageDescribeCommandPresentation,
+  imageDescribeExecutionDefinition,
+} from './semanticIntents/imageDescribe.ts'
 
 interface GeneratedSchemaField extends IntentFieldSpec {
   description?: string
@@ -42,49 +53,14 @@ interface ResolvedIntentNoneInput {
 
 type ResolvedIntentInput = ResolvedIntentLocalFilesInput | ResolvedIntentNoneInput
 
-interface ResolvedIntentSchemaSpec {
-  schema: ZodObject<ZodRawShape>
-}
+type IntentBaseClass =
+  | typeof GeneratedBundledFileIntentCommand
+  | typeof GeneratedNoInputIntentCommand
+  | typeof GeneratedStandardFileIntentCommand
+  | typeof GeneratedWatchableFileIntentCommand
 
-interface ResolvedIntentSingleStepExecution {
-  fixedValues: Record<string, unknown>
-  kind: 'single-step'
-  resultStepName: string
-}
-
-interface ResolvedIntentDynamicExecution {
-  fields: GeneratedSchemaField[]
-  handler: 'image-describe'
-  kind: 'dynamic-step'
-  resultStepName: string
-}
-
-interface ResolvedIntentTemplateExecution {
-  kind: 'template'
-  templateId: string
-}
-
-type ResolvedIntentExecution =
-  | ResolvedIntentDynamicExecution
-  | ResolvedIntentSingleStepExecution
-  | ResolvedIntentTemplateExecution
-
-type ResolvedIntentRunnerKind = 'bundled' | 'no-input' | 'standard' | 'watchable'
-
-interface ResolvedIntentCommandSpec {
-  className: string
-  commandLabel: string
-  description: string
-  details: string
-  examples: Array<[string, string]>
-  execution: ResolvedIntentExecution
-  fieldSpecs: GeneratedSchemaField[]
-  input: ResolvedIntentInput
-  outputDescription: string
-  outputMode?: IntentOutputMode
-  paths: string[]
-  runnerKind: ResolvedIntentRunnerKind
-  schemaSpec?: ResolvedIntentSchemaSpec
+type BuiltIntentCommandDefinition = IntentCommandDefinition & {
+  intentDefinition: IntentFileCommandDefinition | IntentNoInputCommandDefinition
 }
 
 const hiddenFieldNames = new Set([
@@ -99,12 +75,6 @@ const hiddenFieldNames = new Set([
   'stack',
   'use',
 ])
-
-type IntentBaseClass =
-  | typeof GeneratedBundledFileIntentCommand
-  | typeof GeneratedNoInputIntentCommand
-  | typeof GeneratedStandardFileIntentCommand
-  | typeof GeneratedWatchableFileIntentCommand
 
 function toCamelCase(value: string): string {
   return value.replace(/_([a-z])/g, (_match, letter: string) => letter.toUpperCase())
@@ -358,11 +328,11 @@ function collectSchemaFields(
 }
 
 function inferExamples(
-  spec: ResolvedIntentCommandSpec,
+  spec: BuiltIntentCommandDefinition,
   definition?: RobotIntentDefinition,
 ): Array<[string, string]> {
   if (definition == null) {
-    if (spec.execution.kind === 'dynamic-step') {
+    if (spec.intentDefinition.execution.kind === 'dynamic-step') {
       return spec.examples
     }
 
@@ -386,7 +356,12 @@ function inferExamples(
     parts.push('--prompt', JSON.stringify(inferPromptExample(spec.paths)))
   }
 
-  for (const fieldSpec of spec.fieldSpecs) {
+  const fieldSpecs =
+    spec.intentDefinition.execution.kind === 'single-step'
+      ? (spec.intentDefinition.execution.fields as readonly GeneratedSchemaField[])
+      : []
+
+  for (const fieldSpec of fieldSpecs) {
     if (!fieldSpec.required) continue
     if (fieldSpec.name === 'prompt' && inputMode === 'none') continue
 
@@ -395,13 +370,13 @@ function inferExamples(
     parts.push(fieldSpec.optionFlags, exampleValue)
   }
 
-  const outputMode = spec.outputMode ?? 'file'
-  parts.push('--out', inferOutputPath(spec.paths, outputMode, spec.fieldSpecs))
+  const outputMode = spec.intentDefinition.outputMode ?? 'file'
+  parts.push('--out', inferOutputPath(spec.paths, outputMode, fieldSpecs))
 
   return [['Run the command', parts.join(' ')]]
 }
 
-function resolveRobotIntent(definition: RobotIntentDefinition): ResolvedIntentCommandSpec {
+function resolveRobotIntent(definition: RobotIntentDefinition): BuiltIntentCommandDefinition {
   const paths = getIntentPaths(definition)
   const className = `${toPascalCase(paths)}Command`
   const commandLabel = paths.join(' ')
@@ -412,10 +387,20 @@ function resolveRobotIntent(definition: RobotIntentDefinition): ResolvedIntentCo
   const fixedValues = inferFixedValues(definition, input, inputMode)
   const fieldSpecs = collectSchemaFields(schemaShape, fixedValues, input)
   const outputMode = definition.outputMode ?? 'file'
+  const execution: IntentSingleStepExecutionDefinition = {
+    kind: 'single-step',
+    schema,
+    fields: fieldSpecs,
+    fixedValues,
+    resultStepName:
+      getIntentResultStepName(definition) ??
+      (() => {
+        throw new Error(`Could not infer result step name for "${definition.robot}"`)
+      })(),
+  }
 
-  const spec: ResolvedIntentCommandSpec = {
+  const spec: BuiltIntentCommandDefinition = {
     className,
-    commandLabel,
     description: stripTrailingPunctuation(definition.meta.title),
     details:
       inputMode === 'none'
@@ -426,28 +411,26 @@ function resolveRobotIntent(definition: RobotIntentDefinition): ResolvedIntentCo
             ? `Runs \`${definition.robot}\` on each input file and writes the results to \`--out\`.`
             : `Runs \`${definition.robot}\` on each input file and writes the result to \`--out\`.`,
     examples: [],
-    execution: {
-      kind: 'single-step',
-      fixedValues,
-      resultStepName:
-        getIntentResultStepName(definition) ??
-        (() => {
-          throw new Error(`Could not infer result step name for "${definition.robot}"`)
-        })(),
-    },
-    fieldSpecs,
-    input,
-    outputDescription:
-      outputMode === 'directory'
-        ? 'Write the results to this directory'
-        : inputMode === 'local-files'
-          ? 'Write the result to this path or directory'
-          : 'Write the result to this path',
-    outputMode,
     paths,
     runnerKind:
       input.kind === 'none' ? 'no-input' : input.defaultSingleAssembly ? 'bundled' : 'standard',
-    schemaSpec: { schema },
+    intentDefinition:
+      input.kind === 'none'
+        ? {
+            execution,
+            outputDescription: 'Write the result to this path',
+            outputMode,
+          }
+        : {
+            commandLabel,
+            execution,
+            inputPolicy: input.inputPolicy,
+            outputDescription:
+              outputMode === 'directory'
+                ? 'Write the results to this directory'
+                : 'Write the result to this path or directory',
+            outputMode,
+          },
   }
 
   return {
@@ -458,99 +441,50 @@ function resolveRobotIntent(definition: RobotIntentDefinition): ResolvedIntentCo
 
 function resolveImageDescribeIntent(
   definition: SemanticIntentDefinition,
-): ResolvedIntentCommandSpec {
+): BuiltIntentCommandDefinition {
   const paths = getIntentPaths(definition)
+
   return {
     className: `${toPascalCase(paths)}Command`,
-    commandLabel: paths.join(' '),
-    description: 'Describe images as labels or publishable text fields',
-    details:
-      'Generates image labels through `/image/describe`, or structured altText/title/caption/description through `/ai/chat`, then writes the JSON result to `--out`.',
-    examples: [
-      [
-        'Describe an image as labels',
-        'transloadit image describe --input hero.jpg --out labels.json',
-      ],
-      [
-        'Generate WordPress-ready fields',
-        'transloadit image describe --input hero.jpg --for wordpress --out fields.json',
-      ],
-      [
-        'Request a custom field set',
-        'transloadit image describe --input hero.jpg --fields altText,title,caption --out fields.json',
-      ],
-    ],
-    execution: {
-      kind: 'dynamic-step',
-      handler: 'image-describe',
-      resultStepName: 'describe',
-      fields: [
-        {
-          name: 'fields',
-          kind: 'string-array',
-          propertyName: 'fields',
-          optionFlags: '--fields',
-          description:
-            'Describe output fields to generate, for example labels or altText,title,caption,description',
-          required: false,
-        },
-        {
-          name: 'forProfile',
-          kind: 'string',
-          propertyName: 'forProfile',
-          optionFlags: '--for',
-          description: 'Use a named output profile, currently: wordpress',
-          required: false,
-        },
-        {
-          name: 'model',
-          kind: 'string',
-          propertyName: 'model',
-          optionFlags: '--model',
-          description:
-            'Model to use for generated text fields (default: anthropic/claude-sonnet-4-5)',
-          required: false,
-        },
-      ],
-    },
-    fieldSpecs: [],
-    input: {
-      kind: 'local-files',
-      inputPolicy: { kind: 'required' },
-    },
-    outputDescription: 'Write the JSON result to this path or directory',
+    description: imageDescribeCommandPresentation.description,
+    details: imageDescribeCommandPresentation.details,
+    examples: [...imageDescribeCommandPresentation.examples],
     paths,
     runnerKind: 'watchable',
+    intentDefinition: {
+      commandLabel: paths.join(' '),
+      execution: imageDescribeExecutionDefinition,
+      inputPolicy: { kind: 'required' },
+      outputDescription: 'Write the JSON result to this path or directory',
+    },
   }
 }
 
 function resolveTemplateIntent(
   definition: IntentDefinition & { kind: 'template' },
-): ResolvedIntentCommandSpec {
+): BuiltIntentCommandDefinition {
   const outputMode = definition.outputMode ?? 'file'
   const paths = getIntentPaths(definition)
-  const spec: ResolvedIntentCommandSpec = {
+  const spec: BuiltIntentCommandDefinition = {
     className: `${toPascalCase(paths)}Command`,
-    commandLabel: paths.join(' '),
     description: `Run ${stripTrailingPunctuation(definition.templateId)}`,
     details: `Runs the \`${definition.templateId}\` template and writes the outputs to \`--out\`.`,
     examples: [],
-    execution: {
-      kind: 'template',
-      templateId: definition.templateId,
-    },
-    fieldSpecs: [],
-    input: {
-      kind: 'local-files',
-      inputPolicy: { kind: 'required' },
-    },
-    outputDescription:
-      outputMode === 'directory'
-        ? 'Write the results to this directory'
-        : 'Write the result to this path or directory',
-    outputMode,
     paths,
     runnerKind: 'standard',
+    intentDefinition: {
+      commandLabel: paths.join(' '),
+      execution: {
+        kind: 'template',
+        templateId: definition.templateId,
+      },
+      inputPolicy: { kind: 'required' },
+      outputDescription:
+        outputMode === 'directory'
+          ? 'Write the results to this directory'
+          : 'Write the result to this path or directory',
+      outputMode,
+    },
   }
 
   return {
@@ -559,7 +493,7 @@ function resolveTemplateIntent(
   }
 }
 
-function resolveIntent(definition: IntentDefinition): ResolvedIntentCommandSpec {
+function resolveIntent(definition: IntentDefinition): BuiltIntentCommandDefinition {
   if (definition.kind === 'robot') {
     return resolveRobotIntent(definition)
   }
@@ -571,15 +505,7 @@ function resolveIntent(definition: IntentDefinition): ResolvedIntentCommandSpec 
   return resolveTemplateIntent(definition)
 }
 
-function getOptionFields(spec: ResolvedIntentCommandSpec): readonly GeneratedSchemaField[] {
-  if (spec.execution.kind === 'dynamic-step') {
-    return spec.execution.fields
-  }
-
-  return spec.fieldSpecs
-}
-
-function getBaseClass(spec: ResolvedIntentCommandSpec): IntentBaseClass {
+function getBaseClass(spec: BuiltIntentCommandDefinition): IntentBaseClass {
   if (spec.runnerKind === 'no-input') {
     return GeneratedNoInputIntentCommand
   }
@@ -595,7 +521,7 @@ function getBaseClass(spec: ResolvedIntentCommandSpec): IntentBaseClass {
   return GeneratedStandardFileIntentCommand
 }
 
-function createIntentCommandClass(spec: ResolvedIntentCommandSpec): CommandClass {
+function createIntentCommandClass(spec: BuiltIntentCommandDefinition): CommandClass {
   const BaseClass = getBaseClass(spec)
 
   class RuntimeIntentCommand extends BaseClass {}
@@ -606,44 +532,7 @@ function createIntentCommandClass(spec: ResolvedIntentCommandSpec): CommandClass
 
   Object.assign(RuntimeIntentCommand, {
     paths: [spec.paths],
-    intentDefinition:
-      spec.execution.kind === 'single-step'
-        ? {
-            commandLabel: spec.commandLabel,
-            inputPolicy: spec.input.kind === 'local-files' ? spec.input.inputPolicy : undefined,
-            outputDescription: spec.outputDescription,
-            outputMode: spec.outputMode,
-            execution: {
-              kind: 'single-step',
-              schema: spec.schemaSpec?.schema,
-              fields: spec.fieldSpecs,
-              fixedValues: spec.execution.fixedValues,
-              resultStepName: spec.execution.resultStepName,
-            },
-          }
-        : spec.execution.kind === 'dynamic-step'
-          ? {
-              commandLabel: spec.commandLabel,
-              inputPolicy: spec.input.kind === 'local-files' ? spec.input.inputPolicy : undefined,
-              outputDescription: spec.outputDescription,
-              outputMode: spec.outputMode,
-              execution: {
-                kind: 'dynamic-step',
-                handler: spec.execution.handler,
-                fields: spec.execution.fields,
-                resultStepName: spec.execution.resultStepName,
-              },
-            }
-          : {
-              commandLabel: spec.commandLabel,
-              inputPolicy: spec.input.kind === 'local-files' ? spec.input.inputPolicy : undefined,
-              outputDescription: spec.outputDescription,
-              outputMode: spec.outputMode,
-              execution: {
-                kind: 'template',
-                templateId: spec.execution.templateId,
-              },
-            },
+    intentDefinition: spec.intentDefinition,
     usage: Command.Usage({
       category: 'Intent Commands',
       description: spec.description,
@@ -652,7 +541,7 @@ function createIntentCommandClass(spec: ResolvedIntentCommandSpec): CommandClass
     }),
   })
 
-  for (const field of getOptionFields(spec)) {
+  for (const field of getIntentOptionDefinitions(spec.intentDefinition)) {
     Object.defineProperty(RuntimeIntentCommand.prototype, field.propertyName, {
       configurable: true,
       enumerable: true,
